pyproject.toml
[project]
name = "rag-events-assistant"
version = "0.1.0"
description = "RAG-based intelligent events discovery chatbot using Open Agenda API"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Pierre", email = "pierre.pluton@outlook.fr"}
]
keywords = ["rag", "llm", "chatbot", "events", "openagenda", "mistral", "langchain"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core RAG
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-mistralai>=0.2.0",
    # Vector Store
    "faiss-cpu>=1.7.4,<2.0.0",
    "sentence-transformers>=2.3.1,<3.0.0",
    # Interface
    "streamlit>=1.29.0,<2.0.0",
    # Data Processing
    "requests>=2.31.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "pandas>=2.1.4,<3.0.0",
    "beautifulsoup4>=4.12.2,<5.0.0",
    "lxml>=5.0.0,<6.0.0",
    # Configuration & Validation
    "pydantic>=2.5.3,<3.0.0",
    "pydantic-settings>=2.1.0,<3.0.0",
    # Utilities
    "python-json-logger>=2.0.7,<3.0.0",
    "tqdm>=4.66.0,<5.0.0",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
    "nbformat>=5.10.4",
    "mistralai>=1.0.0",
    "ipywidgets>=8.0.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.3,<8.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",

    # Code Quality
    "black>=23.12.0,<24.0.0",
    "ruff>=0.1.8,<1.0.0",
    "mypy>=1.8.0,<2.0.0",
    "pre-commit>=3.6.0,<4.0.0",

    # Jupyter
    "ipython>=8.18.1,<9.0.0",
    "jupyter>=1.0.0,<2.0.0",
    "notebook>=7.0.0,<8.0.0",
    "jupyterlab>=4.0.0,<5.0.0",

    # Type stubs
    "types-requests>=2.31.0",
    "pandas-stubs>=2.1.0",
]

api = [
    "fastapi>=0.109.0,<1.0.0",
    "uvicorn[standard]>=0.25.0,<1.0.0",
]

evaluation = [
    "ragas>=0.1.0,<1.0.0",
    "datasets>=2.16.0,<3.0.0",
]

all = [
    "rag-events-assistant[dev,api,evaluation]",
]

[project.scripts]
rag-events = "src.rag.chatbot:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--cov=src",
    "--cov-report=html",
    "--cov-report=term-missing",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
plugins = ["pydantic.mypy"]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "langchain.*",
    "langchain_community.*",
    "langchain_mistralai.*",
    "faiss.*",
    "sentence_transformers.*",
    "streamlit.*",
]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[dependency-groups]
dev = [
    "tqdm>=4.67.1",
]
